Update how socket endpoint is rotated
Previously when switching servers it would create a new socket and discard the
old one. However, it seemed the bug we had may have been due to create a new
socket in a different thread? Sometimes it would crash and sometimes it wouldn't.
This commit changes it doesn't create a new socket when rotating, instead it
just calls disconnect then reconnect on the same socket. Seems to stop it from
crashing.

closes #1